# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: Play API
  title: Play API
  version: 1.0.0
servers:
- url: /
paths:
  /liveliness:
    get:
      operationId: getLiveliness
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getLiveliness_200_response"
          description: Service is alive
      summary: Check if the service is alive
      x-accepts:
      - application/json
  /auth/login:
    post:
      operationId: postLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/postLogin_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/postLogin_200_response"
          description: Login successful
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/postLogin_401_response"
          description: Unauthorized
      summary: Login to the service
      x-content-type: application/json
      x-accepts:
      - application/json
  /users/:id:
    get:
      operationId: getUserById
      parameters:
      - explode: false
        in: path
        name: id
        required: false
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getUserById_200_response"
          description: App user
      summary: Get a user by ID
      x-accepts:
      - application/json
components:
  schemas:
    postLogin_200_response:
      example:
        token: token
      properties:
        token:
          type: string
      type: object
    getUserById_200_response:
      example:
        id: 1
        username: admin
      properties:
        id:
          example: 1
          type: integer
        username:
          example: admin
          type: string
      type: object
    getLiveliness_200_response:
      example:
        status: alive
      properties:
        status:
          example: alive
          type: string
      type: object
    postLogin_401_response:
      example:
        message: Unauthorized
      properties:
        message:
          example: Unauthorized
          type: string
      type: object
    postLogin_request:
      properties:
        username:
          example: admin
          type: string
        password:
          example: admin
          type: string
      type: object